+++ mek_data_class/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.1.0
+- Removed `DeepCollectionEquality` export in favor of `DataClass.$equals` method
+
 ## 0.0.1+1
 - Fix example documentation links