+++ mek_data_class_generator/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0
+- Fix `hashcode` method generation
+- Removed `@override` annotation to reduce generated code
+- Now the `_props` method is an iterable to improve performance
+
 ## 0.0.3+1
 - Fix example documentation links